fix: scope rate limiting to expensive endpoints only
Address PR review feedback: move rate limiting from the entire jobsRoute
to individual expensive POST endpoints (jobs, probes, lyrics), give chunk
uploads a higher limit, extract the limiter to a shared module with
periodic cleanup to prevent unbounded memory growth, and enable trust
proxy so req.ip resolves correctly behind reverse proxies.
